quell warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21187 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/clientbase/PointShotStrategy.cxx
@@ -143,7 +143,7 @@ float PointShotStrategy::checkShotHit(const ShotCollider& tank,
 
       // construct relative shot ray:  origin and velocity relative to
       // my tank as a function of time (t=0 is start of the interval).
-      const float diffTime = prevTime - s.start;
+      const float diffTime = (float)(prevTime - s.start);
       Ray relativeRay(Intersect::rayMinusRay(s.ray, diffTime,
                                              tankLastMotion, 0.0f));
